+++ test/s3_test.exs
@@ -334,4 +334,20 @@ defmodule S3Test do
     assert URI.to_string(uri) ==
              "https://examplebucket.s3.amazonaws.com/test.txt?X-Amz-Algorithm=AWS4-HMAC-SHA256&X-Amz-Credential=AKIAIOSFODNN7EXAMPLE%2F20130524%2Fus-east-1%2Fs3%2Faws4_request&X-Amz-Date=20130524T000000Z&X-Amz-Expires=86400&X-Amz-SignedHeaders=host&X-Amz-Signature=aeeed9bbccd4d02ee5c0109b86d86835f995330da4c265957d157751f604d404"
   end
+
+  test "ensure build/1 preserves port" do
+    config = fn options -> Keyword.merge(@config, options) end
+
+    assert {%URI{port: 80}, _headers, _body} =
+             S3.build(config.(method: :head, url: "http://example.com"))
+
+    assert {%URI{port: 443}, _headers, _body} =
+             S3.build(config.(method: :head, url: "https://example.com"))
+
+    assert {%URI{port: 8080}, _headers, _body} =
+             S3.build(config.(method: :head, url: "http://example.com:8080"))
+
+    assert {%URI{port: 8443}, _headers, _body} =
+             S3.build(config.(method: :head, url: "https://example.com:8443"))
+  end
 end