+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- ensure HTTP port is set in the returned `URI.t()` from `build/1`
+
 ## 0.1.0 (2024-03-31)
 
 - cleanup, upadte deps, add some docs